+++ src/types/internal.ts
@@ -1,9 +1,20 @@
 import type { ServerResponse } from 'http';
-import type { HttpPathValidator, HttpRequestHandler } from '.';
+import type { Constructor, HttpPathValidator, HttpRequestHandler, IHttpController, IHttpServer } from '.';
 
 export type GroupedHttpRequestHandlers = {
     [method: string]: RequestHandlerContext[];
 };
+
+export interface IInitControllerMethodActionContext {
+    controller: IHttpController;
+    controllerClass: Constructor<IHttpController>;
+    fullFilePath: string;
+    method: HttpRequestHandler;
+    relativeFilePath: string;
+    server: IHttpServer;
+}
+
+export type InitControllerMethodAction = (context: IInitControllerMethodActionContext) => void;
 
 export interface RequestHandlerContext {
     end: (response: ServerResponse) => void;